requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
openai
python-multipart
python-jose[cryptography]
python-dotenv
google-generativeai